fix: deadlock detection for waitgroups
Previously, we checked if the total number of blocked threads are equivalent to (size of thread queue + 1).
This logic is now incorrect as we are not placing blocked threads back into the thread queue. Hence, the total number
of threads may not be (size of thread queue + 1).

Instead, we place any blocked threads into a separate blocked queue. As such, if there are no more threads in the main queue
we can throw a deadlock error as it indicates that there are no more unblocked threads.